\name{Array-subassignment}

\alias{Array-subassignment}
\alias{Array_subassignment}

\alias{subassign_Array_by_logical_array}
\alias{subassign_Array_by_logical_array,Array-method}
\alias{subassign_Array_by_Lindex}
\alias{subassign_Array_by_Lindex,Array-method}
\alias{subassign_Array_by_Mindex}
\alias{subassign_Array_by_Mindex,Array-method}
\alias{subassign_Array_by_Nindex}
\alias{subassign_Array_by_Nindex,Array-method}

\alias{[<-,Array,ANY,ANY,ANY-method}

\title{Low-level generics used in subassignment of Array derivatives}

\description{
  The \pkg{S4Arrays} package defines a small set of low-level generic
  functions to support subassignment of \link{Array} derivatives.
  They are not intended to be used directly by the end user.
}

\seealso{
  \itemize{
    \item The \link{Array} class.
  }
}

\keyword{internal}
\keyword{array}
