\name{Array-subsetting}

\alias{Array-subsetting}
\alias{Array_subsetting}

\alias{subset_Array_by_logical_array}
\alias{subset_Array_by_logical_array,Array-method}
\alias{subset_Array_by_Lindex}
\alias{subset_Array_by_Lindex,Array-method}
\alias{subset_Array_by_Mindex}
\alias{subset_Array_by_Mindex,Array-method}
\alias{subset_Array_by_Nindex}
\alias{subset_Array_by_Nindex,Array-method}

\alias{[,Array,ANY,ANY,ANY-method}

\title{Low-level internal generics involved in subsetting of Array derivatives}

\description{
  The \pkg{S4Arrays} package defines a small set of low-level generic
  functions that are involved in subsetting operations on \link{Array}
  derivatives.
  They are not intended to be used directly by the end user.
}

\seealso{
  \itemize{
    \item The \link{Array} class.
  }
}

\keyword{internal}
\keyword{array}
