\name{readSparseTable}
  
\alias{readSparseTable}

\title{An alternative to \code{scuttle::readSparseCounts}}

\description{
  THIS IS A WORK-IN-PROGRESS!
}

\usage{
readSparseTable(filepath, sep="\t")
}

\arguments{
  \item{filepath}{
    COMING SOON...
  }
  \item{sep}{
    COMING SOON...
  }
}

\details{
  COMING SOON...
}

\value{
  COMING SOON...
}

\seealso{
  \itemize{
    \item \code{\link[scuttle]{readSparseCounts}} in the \pkg{scuttle} package.
  }
}

\examples{
library(scuttle)

example(readSparseCounts)  # to create 'outfile'

readSparseTable(outfile)
}
\keyword{utilities}
